+++ src/core/Separator.js
@@ -1,0 +1,13 @@
+import { Element } from "./Element.js";
+
+export class Separator extends Element {
+  constructor() {
+    super();
+  }
+
+  render() {
+    return `
+            <hr>
+        `.trim();
+  }
+}